+++ pages/stats.tsx
@@ -5,10 +5,10 @@ import {
 	getQuizzesByCategories,
 	getTotalQuiz,
 } from '../redux/slices/stats';
-import { IState } from '../redux/slices/users';
 import { Doughnut } from 'react-chartjs-2';
 import StyledStatsContainer from '../styles/StatsStyled';
 import strings from './strings';
+import { IState } from '../types/slices';
 
 const Panel: React.FC = () => {
 	const dispatch = useDispatch();
@@ -25,7 +25,6 @@ const Panel: React.FC = () => {
 	} = useSelector((state: IState) => state.stats);
 	const { language } = useSelector((state: IState) => state.global);
 	const s = strings[language];
-
 
 	if (loading) return <h1>Cargando...</h1>;
 
@@ -54,12 +53,16 @@ const Panel: React.FC = () => {
 			<StyledStatsContainer>
 				<h1 className='title'>{s.statsTitle}</h1>
 				<div className='StatsContainer'>
-					<h1 style={{marginRight: '1em'}}>{s.statsQuantity} {totalQuizzes}</h1>
-					<h1>{s.statsUsers} {totalNewUsers}</h1>
+					<h1 style={{ marginRight: '1em' }}>
+						{s.statsQuantity} {totalQuizzes}
+					</h1>
+					<h1>
+						{s.statsUsers} {totalNewUsers}
+					</h1>
 				</div>
 				<div className='DougContainer'>
 					<Doughnut data={data} width={400} height={300} />
-					</div>
+				</div>
 			</StyledStatsContainer>
 		</main>
 	);
