+++ pages/users.tsx
@@ -1,7 +1,8 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import CardsContainer from '../components/CardsContainer';
-import { getUsers, IState } from '../redux/slices/users';
+import { getUsers } from '../redux/slices/users';
+import { IState } from '../types/slices';
 
 const Panel: React.FC = () => {
 	const dispatch = useDispatch();
